Update database.js
Looks like a misunderstanding of what `pgp.end()` does and why or where to use it. See [Library De-initilization](https://github.com/vitaly-t/pg-promise#library-de-initialization)

And in a context like this one it will do precisely nothing. You are trying to shit down the database connection pool before it is even used for the first time. Best advise - don't use it at all.